Help Center: use a named capturing request client in the CTA test to satisfy Phan
