+++ projects/packages/help-center/tests/php/Cta_Request_Test.php
@@ -6,7 +6,8 @@
  */
 
 use Automattic\Jetpack\Help_Center\WP_REST_Help_Center_CTA;
-use Automattic\Jetpack\Help_Center\Wpcom_Request_Client;
+
+require_once __DIR__ . '/class-cta-capturing-request-client.php';
 
 /**
  * Class Cta_Request_Test
@@ -63,64 +64,7 @@ class Cta_Request_Test extends \WorDBless\BaseTestCase {
 	 * @param int    $code The upstream status code.
 	 * @param string $body The upstream response body.
 	 */
-	private function create_request_client( int $code, string $body ): Wpcom_Request_Client {
-		return new class( $code, $body ) implements Wpcom_Request_Client {
-			/**
-			 * Captured requests.
-			 *
-			 * @var array
-			 */
-			public $requests = array();
-
-			/**
-			 * The upstream status code.
-			 *
-			 * @var int
-			 */
-			private $code;
-
-			/**
-			 * The upstream response body.
-			 *
-			 * @var string
-			 */
-			private $response_body;
-
-			/**
-			 * Constructor.
-			 *
-			 * @param int    $code          The upstream status code.
-			 * @param string $response_body The upstream response body.
-			 */
-			public function __construct( int $code, string $response_body ) {
-				$this->code          = $code;
-				$this->response_body = $response_body;
-			}
-
-			public function is_user_connected() {
-				return true;
-			}
-
-			public function request(
-				$path,
-				$version = '2',
-				$args = array(),
-				$body = null,
-				$base_api_path = 'wpcom'
-			) {
-				$this->requests[] = compact( 'path', 'version', 'args', 'body', 'base_api_path' );
-
-				return array(
-					'headers'  => array(),
-					'body'     => $this->response_body,
-					'response' => array(
-						'code'    => $this->code,
-						'message' => '',
-					),
-					'cookies'  => array(),
-					'filename' => null,
-				);
-			}
-		};
+	private function create_request_client( int $code, string $body ): Cta_Capturing_Request_Client {
+		return new Cta_Capturing_Request_Client( $code, $body );
 	}
 }
